Improve functionality when interacting with live simulations
Some simulations have a frequency of output that is high enough
that the particles are updated in the time between reading the
first and last particles. This has been fixed in a few functions
by fixing the number of indices to store based on the length of
the first particle series.